+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- start at login for Windows Store version
+
+## Changed
+- updated many translations
 
 ## [1.14.1] - 2023-06-04
 ## Added